test: ✅ move test to test dir
