+++ test/domain/services/SQLManagerService.test.ts
